+++ frontend/src/features/setup-course/StepList.js
