+++ frontend/src/features/setup-course/icons.js
